requirements.txt
flask
torch
Pillow
ultralytics
gunicorn
